backend/requirements.txt
fastapi
uvicorn
pydantic>=2.0.0
scikit-learn
google-genai
python-dotenv
requests
